move pipelineCanvasState down to prevent re-rendering PipelineEditor when scaling
